+++ mitra_models/migrations/schema.sql
@@ -16,6 +16,14 @@ CREATE TABLE instance (
     hostname VARCHAR(100) PRIMARY KEY
 );
 
+CREATE TABLE filter_rule (
+    id INTEGER GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    target VARCHAR(2000) NOT NULL,
+    filter_action SMALLINT NOT NULL,
+    is_reversed BOOLEAN NOT NULL,
+    UNIQUE (target, filter_action)
+);
+
 CREATE TABLE actor_profile (
     id UUID PRIMARY KEY,
     user_id UUID UNIQUE,