+++ mitra_validators/src/lib.rs
@@ -2,6 +2,7 @@ pub mod activitypub;
 pub mod custom_feeds;
 pub mod emojis;
 pub mod errors;
+pub mod filter_rules;
 pub mod invoices;
 pub mod media;
 pub mod oauth;